requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
azure-cognitiveservices-speech==1.34.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
elevenlabs==0.2.27
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipython==8.19.0
jedi==0.19.1
lxml==5.0.1
matplotlib-inline==0.1.6
numpy==1.26.3
openai==1.6.1
pandas==2.1.4
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
Pygments==2.17.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
spotipy==2.23.0
stack-data==0.6.3
tiktoken==0.5.2
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.7
urllib3==2.1.0
wcwidth==0.2.13
websockets==12.0
